Performs some suspicious, usually malware-related, activities within a Windows system.
This PE file is designed to perform some suspicious, usually malware-related, activities within a Windows system. It will set off various signatures inside a sandbox environment. This way the operation of a sandbox can be tested via a safe file, that is performing documented and verifyable tasks (source code is available here).

https://www.hybrid-analysis.com/sample/4913962947d01b0fbf4275839c001af705c0dae07fa859e76cfc007f91fc68ed/5b6a72557ca3e13bc340be83